refactor(useDefindex): extract cache/validation utilities to lib/defindex/vault-cache.ts
Step 1/4 of #66. Moves readErrorMessage, hasClientVaultConfigured,
vaultMetaRequest, vaultBalanceRequest, sacBalanceRequest,
invalidateVaultDataCache, isRateLimitMessage, and their supporting types
out of useDefindex.ts into a dedicated cache module. No behavior change —
useDefindex.ts now imports these instead of defining them locally.

Verified: tsc --noEmit error count unchanged (10, same as baseline).
No consumer files modified.
